# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Радар обязанностей — API мини-приложения",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/company": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Карточка подключённой компании",
        "operationId": "current_company_api_company_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/session": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Подключённая компания; 404 — компания ещё не подключена",
        "operationId": "current_company_api_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "company"
        ],
        "summary": "Подключить компанию по ИНН: профиль из реестра МСП, сроки обязанностей; 422 — неверный ИНН, 404 — нет в реестре МСП, 409 — ведёт другой пользователь, 503 — реестр ФНС не ответил",
        "operationId": "connect_company_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/company/refresh": {
      "post": {
        "tags": [
          "company"
        ],
        "summary": "Перепроверить компанию в реестре МСП сейчас; ЕГРЮЛ и ЕРКНМ — в фоне",
        "operationId": "refresh_company_api_company_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/company/profile": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Экран «Данные компании»: профиль, данные реестра и варианты ответов",
        "operationId": "profile_form_api_company_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileForm"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "company"
        ],
        "summary": "Сохранить режим, численность и поправки к реестру; обязанности пересчитываются",
        "operationId": "save_profile_api_company_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/settings/notifications": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Настройки уведомлений",
        "operationId": "get_notification_settings_api_settings_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Сохранить настройки уведомлений",
        "operationId": "save_notification_settings_api_settings_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/settings/notifications/seen": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Лента просмотрена — гасит отметку непрочитанного",
        "operationId": "mark_notifications_seen_api_settings_notifications_seen_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Benefit": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "basis": {
            "type": "string",
            "title": "Basis"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "title",
          "basis",
          "url"
        ],
        "title": "Benefit"
      },
      "Company": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "initials": {
            "type": "string",
            "title": "Initials"
          },
          "regime": {
            "type": "string",
            "title": "Regime"
          },
          "inn": {
            "type": "string",
            "title": "Inn"
          },
          "kpp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kpp"
          },
          "ogrn": {
            "type": "string",
            "title": "Ogrn"
          },
          "okved": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Okved"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "headcount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headcount"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "needsAnswers": {
            "type": "boolean",
            "title": "Needsanswers"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "benefits": {
            "items": {
              "$ref": "#/components/schemas/Benefit"
            },
            "type": "array",
            "title": "Benefits",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "fullName",
          "initials",
          "regime",
          "inn",
          "ogrn",
          "needsAnswers",
          "source"
        ],
        "title": "Company"
      },
      "ConnectRequest": {
        "properties": {
          "inn": {
            "type": "string",
            "pattern": "^(\\d{10}|\\d{12})$",
            "title": "Inn"
          }
        },
        "type": "object",
        "required": [
          "inn"
        ],
        "title": "ConnectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationSettings": {
        "properties": {
          "chat": {
            "type": "boolean",
            "title": "Chat",
            "default": true
          },
          "push": {
            "type": "boolean",
            "title": "Push",
            "default": false
          },
          "email": {
            "type": "boolean",
            "title": "Email",
            "default": false
          },
          "emailAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emailaddress"
          },
          "quiet": {
            "type": "boolean",
            "title": "Quiet",
            "default": true
          },
          "quietRange": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}–\\d{2}:\\d{2}$",
            "title": "Quietrange",
            "default": "22:00–08:00"
          },
          "remind": {
            "type": "string",
            "enum": [
              "d30-7-1",
              "d7-3-1",
              "d3-0",
              "d0"
            ],
            "title": "Remind",
            "default": "d30-7-1"
          }
        },
        "type": "object",
        "title": "NotificationSettings"
      },
      "Option": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "title": "Option"
      },
      "ProfileForm": {
        "properties": {
          "isLegalEntity": {
            "type": "boolean",
            "title": "Islegalentity"
          },
          "regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime"
          },
          "headcount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headcount"
          },
          "registryHeadcount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registryheadcount"
          },
          "okved": {
            "type": "string",
            "title": "Okved"
          },
          "registryOkved": {
            "type": "string",
            "title": "Registryokved"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "registryRegion": {
            "type": "string",
            "title": "Registryregion"
          },
          "hasLicenses": {
            "type": "boolean",
            "title": "Haslicenses"
          },
          "flags": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Flags"
          },
          "patentFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patentfrom"
          },
          "patentTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patentto"
          },
          "options": {
            "$ref": "#/components/schemas/ProfileOptions"
          }
        },
        "type": "object",
        "required": [
          "isLegalEntity",
          "okved",
          "registryOkved",
          "region",
          "registryRegion",
          "hasLicenses",
          "flags",
          "options"
        ],
        "title": "ProfileForm"
      },
      "ProfileOptions": {
        "properties": {
          "regimes": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Regimes"
          },
          "headcounts": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Headcounts"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Regions"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "regimes",
          "headcounts",
          "regions",
          "flags"
        ],
        "title": "ProfileOptions"
      },
      "ProfileUpdate": {
        "properties": {
          "regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime"
          },
          "headcount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headcount"
          },
          "okved": {
            "type": "string",
            "title": "Okved"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "hasLicenses": {
            "type": "boolean",
            "title": "Haslicenses"
          },
          "flags": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Flags",
            "default": {}
          },
          "patentFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patentfrom"
          },
          "patentTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patentto"
          }
        },
        "type": "object",
        "required": [
          "okved",
          "region",
          "hasLicenses"
        ],
        "title": "ProfileUpdate"
      },
      "Source": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "name",
          "demo",
          "updatedAt"
        ],
        "title": "Source"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}